refactor: 이슈검증 리펙터링 및 updateIssue 메서드 구현 시작
