+++ src/main/java/dynamicquad/agilehub/issue/service/IssueService.java
@@ -1,11 +1,10 @@
 package dynamicquad.agilehub.issue.service;
 
-import dynamicquad.agilehub.global.exception.GeneralException;
-import dynamicquad.agilehub.global.header.status.ErrorStatus;
 import dynamicquad.agilehub.issue.controller.request.IssueRequest.IssueCreateRequest;
+import dynamicquad.agilehub.issue.domain.Issue;
 import dynamicquad.agilehub.issue.service.factory.IssueFactoryProvider;
 import dynamicquad.agilehub.project.domain.Project;
-import dynamicquad.agilehub.project.domain.ProjectRepository;
+import dynamicquad.agilehub.project.service.ProjectValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,20 +14,28 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class IssueService {
 
-    private final IssueValidator issueValidator;
+    private final ProjectValidator projectValidator;
     private final IssueFactoryProvider issueFactoryProvider;
+    private final IssueValidator issueValidator;
 
     @Transactional
     public Long createIssue(String key, IssueCreateRequest request) {
-        Project project = issueValidator.findProject(key);
+        Project project = projectValidator.findProject(key);
 
         return issueFactoryProvider.getIssueFactory(request.getType())
             .createIssue(request, project);
 
     }
 
-    private Project findProject(String key) {
-        return projectRepository.findByKey(key)
-            .orElseThrow(() -> new GeneralException(ErrorStatus.PROJECT_NOT_FOUND));
+    @Transactional
+    public void updateIssue(String key, Long issueId, IssueCreateRequest request) {
+        Project project = projectValidator.findProject(key);
+        Issue issue = issueValidator.findIssue(issueId);
+        issueValidator.validateIssueInProject(project, issue);
+
+        issueFactoryProvider.getIssueFactory(request.getType())
+            .updateIssue(issue, project, request);
     }
+
+
 }
