+++ src/main/java/dynamicquad/agilehub/issue/service/IssueValidator.java
@@ -3,11 +3,8 @@ package dynamicquad.agilehub.issue.service;
 import dynamicquad.agilehub.global.exception.GeneralException;
 import dynamicquad.agilehub.global.header.status.ErrorStatus;
 import dynamicquad.agilehub.issue.domain.Issue;
-import dynamicquad.agilehub.member.domain.Member;
-import dynamicquad.agilehub.member.domain.MemberRepository;
-import dynamicquad.agilehub.project.domain.MemberProjectRepository;
+import dynamicquad.agilehub.issue.domain.IssueRepository;
 import dynamicquad.agilehub.project.domain.Project;
-import dynamicquad.agilehub.project.domain.ProjectRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,29 +12,7 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class IssueValidator {
 
-    private final ProjectRepository projectRepository;
-    private final MemberRepository memberRepository;
-    private final MemberProjectRepository memberProjectRepository;
-
-    public Project findProject(String key) {
-        return projectRepository.findByKey(key)
-            .orElseThrow(() -> new GeneralException(ErrorStatus.PROJECT_NOT_FOUND));
-    }
-
-    public Member findMember(Long assigneeId, Long projectId) {
-
-        if (assigneeId == null) {
-            return null;
-        }
-
-        Member member = memberRepository.findById(assigneeId)
-            .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
-
-        memberProjectRepository.findByMemberIdAndProjectId(member.getId(), projectId)
-            .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_IN_PROJECT));
-
-        return member;
-    }
+    private final IssueRepository issueRepository;
 
     public void validateIssueInProject(Project project, Issue issue) {
         if (!issue.getProject().getId().equals(project.getId())) {
@@ -45,5 +20,9 @@ public class IssueValidator {
         }
     }
 
+    public Issue findIssue(Long issueId) {
+        return issueRepository.findById(issueId)
+            .orElseThrow(() -> new GeneralException(ErrorStatus.ISSUE_NOT_FOUND));
+    }
 
 }
