+++ src/main/java/dynamicquad/agilehub/project/service/ProjectService.java
@@ -17,6 +17,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class ProjectService {
 
+    private final ProjectValidator projectValidator;
     private final ProjectRepository projectRepository;
 
     @Transactional
@@ -32,7 +33,7 @@ public class ProjectService {
     @Transactional
     public String updateProject(String originKey, ProjectUpdateRequest request) {
 
-        Project project = findProject(originKey);
+        Project project = projectValidator.findProject(originKey);
 
         validateKeyUniqueness(request.getKey());
 
@@ -40,10 +41,6 @@ public class ProjectService {
 
     }
 
-    private Project findProject(String originKey) {
-        return projectRepository.findByKey(originKey)
-            .orElseThrow(() -> new GeneralException(ErrorStatus.PROJECT_NOT_FOUND));
-    }
 
     private void validateKeyUniqueness(String key) {
         if (projectRepository.existsByKey(key)) {
